Solution using the all() method alongside the search() method.
This uses the all() method which searches through an iterable, in this case the list passwordElements is the iterable.

all() checks to ensure that every iteration is true and then returns True.  This uses the multiple arguments from the search() method.

re.search(x, y) will search the string in y for x and return the first finding.  Here, r is drawn from the passwordElements list.